+++ src/dotnet/Fable.Compiler/Transforms/Fable2Babel.fs
@@ -181,16 +181,16 @@ module Util =
         CallExpression(get None (Identifier "Object") methodName, args) :> Expression
 
     let coreUtil (com: IBabelCompiler) ctx memberName args =
-        CallExpression(com.TransformImport(ctx, memberName, "Util", Fable.CoreLib), args) :> Expression
+        CallExpression(com.TransformImport(ctx, memberName, "Util", Fable.Precompiled), args) :> Expression
 
     let coreLibCall (com: IBabelCompiler) ctx moduleName memberName args =
-        CallExpression(com.TransformImport(ctx, memberName, moduleName, Fable.CoreLib), args) :> Expression
+        CallExpression(com.TransformImport(ctx, memberName, moduleName, Fable.Precompiled), args) :> Expression
 
     let coreLibConstructorCall (com: IBabelCompiler) ctx moduleName memberName args =
-        NewExpression(com.TransformImport(ctx, memberName, moduleName, Fable.CoreLib), args) :> Expression
+        NewExpression(com.TransformImport(ctx, memberName, moduleName, Fable.Precompiled), args) :> Expression
 
     let coreValue (com: IBabelCompiler) ctx moduleName memberName =
-        com.TransformImport(ctx, memberName, moduleName, Fable.CoreLib)
+        com.TransformImport(ctx, memberName, moduleName, Fable.Precompiled)
 
     let entityRefMaybeImported (com: IBabelCompiler) ctx ent =
         if FSharp2Fable.Util.isReplacementCandidate ent then
@@ -1576,7 +1576,7 @@ module Compiler =
                     let sanitizedPath =
                         match kind with
                         | Fable.CustomImport | Fable.Internal _ -> path
-                        | Fable.CoreLib -> com.FableCore + "/" + path + Naming.targetFileExtension
+                        | Fable.Precompiled -> com.FablePrecompiled + "/" + path + Naming.targetFileExtension
                     let i =
                       { Selector =
                             if selector = Naming.placeholder
@@ -1599,7 +1599,7 @@ module Compiler =
 
         interface ICompiler with
             member __.Options = com.Options
-            member __.FableCore = com.FableCore
+            member __.FablePrecompiled = com.FablePrecompiled
             member __.CurrentFile = com.CurrentFile
             member __.GetUniqueVar(name) = com.GetUniqueVar(?name=name)
             member __.GetRootModule(fileName) = com.GetRootModule(fileName)