+++ src/js/fable-compiler-js/src/app.fs
@@ -116,10 +116,10 @@ let parseFiles projectPath outDir optimized commonjs =
     let fileNames = fileNames |> Array.filter (fun x -> not (x.EndsWith(".fsi")))
 
     // Fable (F# to Babel)
-    let fableCoreDir = "fable-core"
-    let fableCoreDist = if commonjs then "/fable-core-commonjs" else "/bundle/fable-core"
-    copyFolder (__dirname + fableCoreDist, Path.Combine(outDir, fableCoreDir))
-    let parseFable (fileName, ast) = fable.CompileToBabelAst(fableCoreDir, ast, fileName, optimized)
+    let fablePrecompiledDir = "fable-precompiled"
+    let fablePrecompiledDist = if commonjs then "/fable-precompiled-commonjs" else "/bundle/fable-precompiled"
+    copyFolder (__dirname + fablePrecompiledDist, Path.Combine(outDir, fablePrecompiledDir))
+    let parseFable (fileName, ast) = fable.CompileToBabelAst(fablePrecompiledDir, ast, fileName, optimized)
     let fsAst = parseRes.ProjectResults
     for fileName in fileNames do